# server/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["Root"]}},"/booking/{bookingId}":{"get":{"operationId":"BookingController_getBookingById","parameters":[{"name":"bookingId","required":true,"in":"path","description":"The ID of the booking","schema":{"example":"123e4567-e89b-12d3-a456-426614174000","type":"string"}}],"responses":{"200":{"description":"Returns the booking with the given ID","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDto"}}}}},"tags":["Bookings"]}},"/booking/availability/{parkId}/{date}/{duration}":{"get":{"operationId":"BookingController_getParkAvailabilityByDate","parameters":[{"name":"parkId","required":true,"in":"path","description":"The ID of the park","schema":{"type":"number"}},{"name":"date","required":true,"in":"path","description":"Date to check availability for formatted as YYYY-MM-DD","schema":{"type":"string"}},{"name":"duration","required":true,"in":"path","description":"The duration of the booking in minutes: must be 30, 60, 90, 120","schema":{"enum":[30,60,90,120],"type":"number"}}],"responses":{"200":{"description":"Returns the availability for the park on the given date","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParkAvailabilityDto"}}}}},"tags":["Bookings"]}},"/booking/reserve":{"post":{"operationId":"BookingController_requestBooking","parameters":[],"requestBody":{"required":true,"description":"The booking to reserve","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestBookingDto"}}}},"responses":{"201":{"description":"The booking has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDto"}}}}},"tags":["Bookings"]}},"/booking/confirm":{"post":{"operationId":"BookingController_confirmBooking","parameters":[],"requestBody":{"required":true,"description":"Booking ID and email of the user confirming the booking","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModifyBookingDto"}}}},"responses":{"201":{"description":"The booking has been successfully confirmed.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingConfirmationDto"}}}}},"tags":["Bookings"]}},"/booking/cancel":{"post":{"operationId":"BookingController_confirmCancellation","parameters":[],"requestBody":{"required":true,"description":"Booking ID and email of the user canceling the booking","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModifyBookingDto"}}}},"responses":{"201":{"description":"The booking has been successfully canceled.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingConfirmationDto"}}}}},"tags":["Bookings"]}},"/court":{"get":{"operationId":"CourtController_getCourts","parameters":[],"responses":{"200":{"description":"Returns all courts","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Court"}}}}}},"tags":["Courts"]}},"/park":{"get":{"operationId":"ParkController_getAllParks","parameters":[],"responses":{"200":{"description":"Returns all parks","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ParkDto"}}}}}},"tags":["Parks"]},"post":{"operationId":"ParkController_createPark","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateParkDto"}}}},"responses":{"201":{"description":"The park has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParkDto"}}}}},"tags":["Parks"]}},"/park/{id}":{"get":{"operationId":"ParkController_getPark","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Retrieves a Park and it's related courts by Id","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ParkDto"}}}}}},"tags":["Parks"]}},"/user":{"get":{"operationId":"UserController_getAllUsers","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/user/{id}":{"get":{"operationId":"UserController_getUserById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}}},"info":{"title":"GameSetBook API","description":"API for Booking Tennis and Pickleball Courts","version":"1.0","contact":{}},"tags":[{"name":"GameSetBook","description":""}],"servers":[],"components":{"schemas":{"BookingDto":{"type":"object","properties":{"id":{"type":"string","description":"The ID of the booking","example":"123e4567-e89b-12d3-a456-426614174000"},"courtId":{"type":"number","description":"The ID of the court being booked","example":1},"parkId":{"type":"number","description":"The ID of the park being booked","example":1},"timeConfirmed":{"type":"string","description":"The time the booking was confirmed","example":"2024-01-01T11:00:00.000Z"},"start":{"type":"string","description":"The start time of the booking","example":"2024-01-01T11:00:00.000Z"},"end":{"type":"string","description":"The end time of the booking","example":"2024-01-01T12:00:00.000Z"},"duration":{"type":"number","description":"The duration of the booking in minutes","example":"60","enum":[30,60,90,120]},"status":{"type":"number","description":"The status of the booking","enum":[1,2,3]},"userId":{"type":"string","description":"The ID of the user who made the booking","example":"123e4567-e89b-12d3-a456-426614174000"},"timezone":{"type":"string","description":"The timezone of the park","enum":["America/New_York","America/Chicago","America/Denver","America/Phoenix","America/Los_Angeles","America/Indiana/Indianapolis","America/Indiana/Knox","America/Indiana/Winamac","America/Kentucky/Louisville","America/Kentucky/Monticello","America/Detroit","America/North_Dakota/Center","America/North_Dakota/New_Salem","America/North_Dakota/Beulah"]}},"required":["id","courtId","parkId","start","end","duration","status","userId","timezone"]},"BookingDurationDisplayDto":{"type":"object","properties":{"minutes":{"type":"number","description":"The duration of the slot.","example":"30","enum":[30,60,90,120]},"display":{"type":"string","description":"The display string for the duration.","example":"30 minutes"}},"required":["minutes","display"]},"CourtDto":{"type":"object","properties":{"courtNumber":{"type":"number","description":"The court number within a park.","example":1},"timezone":{"type":"string","description":"The timezone for the park.","example":"America/New_York","enum":["America/New_York","America/Chicago","America/Denver","America/Phoenix","America/Los_Angeles","America/Indiana/Indianapolis","America/Indiana/Knox","America/Indiana/Winamac","America/Kentucky/Louisville","America/Kentucky/Monticello","America/Detroit","America/North_Dakota/Center","America/North_Dakota/New_Salem","America/North_Dakota/Beulah"]},"type":{"type":"number","description":"Tennis or Pickleball.","example":"Tennis","enum":[1,2]},"configuration":{"type":"number","description":"How the court is configured with respect to Tennis and Pickleball.","example":"Tennis","enum":[1,2,3,4]},"compositeId":{"type":"string","description":"Unique identifier for composite courts, meaning a single Tennis Court with multiple uses or multiple Pickleball courts on a single Tennis court.","example":"123e4567-e89b-12d3-a456-426614174000"},"id":{"type":"number","description":"The ID of the court","example":1}},"required":["courtNumber","timezone","type","configuration","id"]},"ParkAvailabilitySlot":{"type":"object","properties":{"start":{"type":"string","description":"The start time of the slot.","example":"2024-01-01T00:00:00.000Z"},"end":{"type":"string","description":"The end time of the slot.","example":"2024-01-01T00:00:00.000Z"},"duration":{"description":"The duration of the slot.","example":"30","allOf":[{"$ref":"#/components/schemas/BookingDurationDisplayDto"}]},"courts":{"description":"The list of courts available for the slot.","example":[],"type":"array","items":{"$ref":"#/components/schemas/CourtDto"}}},"required":["start","end","duration","courts"]},"ParkAvailabilityDto":{"type":"object","properties":{"parkId":{"type":"number","description":"The ID of the park to get availability for.","example":1},"timezone":{"type":"string","description":"The timezone for the park.","example":"America/New_York","enum":["America/New_York","America/Chicago","America/Denver","America/Phoenix","America/Los_Angeles","America/Indiana/Indianapolis","America/Indiana/Knox","America/Indiana/Winamac","America/Kentucky/Louisville","America/Kentucky/Monticello","America/Detroit","America/North_Dakota/Center","America/North_Dakota/New_Salem","America/North_Dakota/Beulah"]},"slots":{"description":"The list of Slots available for the park on a given date.","example":[],"type":"array","items":{"$ref":"#/components/schemas/ParkAvailabilitySlot"}}},"required":["parkId","timezone","slots"]},"RequestBookingDto":{"type":"object","properties":{"parkId":{"type":"number","description":"The ID of the park being booked","example":1},"courtId":{"type":"number","description":"The ID of the court being booked","example":1},"timezone":{"type":"string","description":"The timezone of the court being booked","example":"America/New_York","enum":["America/New_York","America/Chicago","America/Denver","America/Phoenix","America/Los_Angeles","America/Indiana/Indianapolis","America/Indiana/Knox","America/Indiana/Winamac","America/Kentucky/Louisville","America/Kentucky/Monticello","America/Detroit","America/North_Dakota/Center","America/North_Dakota/New_Salem","America/North_Dakota/Beulah"]},"email":{"type":"string","description":"The email of the user booking the court","example":"test@test.com"},"start":{"type":"string","description":"The start time of the booking","example":"2024-01-01T11:00:00.000Z"},"end":{"type":"string","description":"The end time of the booking","example":"2024-01-01T12:00:00.000Z"},"duration":{"type":"number","description":"The duration of the booking in minutes","example":"30, 60, 90, 120","enum":[30,60,90,120]}},"required":["parkId","courtId","timezone","email","start","end","duration"]},"ModifyBookingDto":{"type":"object","properties":{"bookingId":{"type":"string","description":"The ID of the booking to confirm","example":"123e4567-e89b-12d3-a456-426614174000"},"email":{"type":"string","description":"The email of the user confirming the booking","example":"test@test.com"}},"required":["bookingId","email"]},"BookingConfirmationDto":{"type":"object","properties":{"id":{"type":"string","description":"The ID of the booking","example":"123e4567-e89b-12d3-a456-426614174000"},"courtId":{"type":"number","description":"The ID of the court being booked","example":1},"parkId":{"type":"number","description":"The ID of the park being booked","example":1},"timeConfirmed":{"type":"string","description":"The time the booking was confirmed","example":"2024-01-01T11:00:00.000Z"},"start":{"type":"string","description":"The start time of the booking","example":"2024-01-01T11:00:00.000Z"},"end":{"type":"string","description":"The end time of the booking","example":"2024-01-01T12:00:00.000Z"},"duration":{"type":"number","description":"The duration of the booking in minutes","example":"60","enum":[30,60,90,120]},"status":{"type":"number","description":"The status of the booking","enum":[1,2,3]},"userId":{"type":"string","description":"The ID of the user who made the booking","example":"123e4567-e89b-12d3-a456-426614174000"},"timezone":{"type":"string","description":"The timezone of the park","enum":["America/New_York","America/Chicago","America/Denver","America/Phoenix","America/Los_Angeles","America/Indiana/Indianapolis","America/Indiana/Knox","America/Indiana/Winamac","America/Kentucky/Louisville","America/Kentucky/Monticello","America/Detroit","America/North_Dakota/Center","America/North_Dakota/New_Salem","America/North_Dakota/Beulah"]},"parkName":{"type":"string","description":"The name of the park","example":"Central Park Tennis Courts"},"fullParkAddress":{"type":"string","description":"The full address of the park","example":"123 Main St, New York, NY 10001"},"courtNumber":{"type":"number","description":"The Number of the court being booked","example":1},"userEmail":{"type":"string","description":"The ID of the user who made the booking","example":"123e4567-e89b-12d3-a456-426614174000"}},"required":["id","courtId","parkId","start","end","duration","status","userId","timezone","parkName","fullParkAddress","courtNumber","userEmail"]},"Court":{"type":"object","properties":{}},"CreateCourtDto":{"type":"object","properties":{"courtNumber":{"type":"number","description":"The court number within a park.","example":1},"timezone":{"type":"string","description":"The timezone for the park.","example":"America/New_York","enum":["America/New_York","America/Chicago","America/Denver","America/Phoenix","America/Los_Angeles","America/Indiana/Indianapolis","America/Indiana/Knox","America/Indiana/Winamac","America/Kentucky/Louisville","America/Kentucky/Monticello","America/Detroit","America/North_Dakota/Center","America/North_Dakota/New_Salem","America/North_Dakota/Beulah"]},"type":{"type":"number","description":"Tennis or Pickleball.","example":"Tennis","enum":[1,2]},"configuration":{"type":"number","description":"How the court is configured with respect to Tennis and Pickleball.","example":"Tennis","enum":[1,2,3,4]},"compositeId":{"type":"string","description":"Unique identifier for composite courts, meaning a single Tennis Court with multiple uses or multiple Pickleball courts on a single Tennis court.","example":"123e4567-e89b-12d3-a456-426614174000"}},"required":["courtNumber","timezone","type","configuration"]},"ParkDto":{"type":"object","properties":{"name":{"type":"string","description":"The name of the park.","example":"The Tennis Club"},"addressLine":{"type":"string","description":"The address line for the park.","example":"123 Main St"},"city":{"type":"string","description":"The city for the park.","example":"San Francisco"},"state":{"type":"string","description":"The state for the park (e.g. CA).","example":"CA"},"zip":{"type":"string","description":"The zip code for the park.","example":"94103"},"timezone":{"type":"string","description":"The timezone for the park.","example":"America/New_York","enum":["America/New_York","America/Chicago","America/Denver","America/Phoenix","America/Los_Angeles","America/Indiana/Indianapolis","America/Indiana/Knox","America/Indiana/Winamac","America/Kentucky/Louisville","America/Kentucky/Monticello","America/Detroit","America/North_Dakota/Center","America/North_Dakota/New_Salem","America/North_Dakota/Beulah"]},"lighted":{"type":"boolean","description":"Whether the park has lights.","example":true},"bathrooms":{"type":"boolean","description":"Whether the park has bathrooms.","example":true},"tennisClub":{"type":"boolean","description":"Whether the park is a tennis club.","example":true},"ballWall":{"type":"boolean","description":"Whether the park has a ball wall.","example":true},"tennisStore":{"type":"boolean","description":"Whether the park has a tennis store.","example":true},"fee":{"type":"boolean","description":"Whether the park has a fee.","example":true},"restricted":{"type":"boolean","description":"Whether the park is restricted.","example":true},"courts":{"description":"The list of courts associated with the park.","example":[],"type":"array","items":{"$ref":"#/components/schemas/CreateCourtDto"}},"id":{"type":"number","description":"The ID of the park","example":1}},"required":["name","addressLine","city","state","zip","timezone","id"]},"CreateParkDto":{"type":"object","properties":{"name":{"type":"string","description":"The name of the park.","example":"The Tennis Club"},"addressLine":{"type":"string","description":"The address line for the park.","example":"123 Main St"},"city":{"type":"string","description":"The city for the park.","example":"San Francisco"},"state":{"type":"string","description":"The state for the park (e.g. CA).","example":"CA"},"zip":{"type":"string","description":"The zip code for the park.","example":"94103"},"timezone":{"type":"string","description":"The timezone for the park.","example":"America/New_York","enum":["America/New_York","America/Chicago","America/Denver","America/Phoenix","America/Los_Angeles","America/Indiana/Indianapolis","America/Indiana/Knox","America/Indiana/Winamac","America/Kentucky/Louisville","America/Kentucky/Monticello","America/Detroit","America/North_Dakota/Center","America/North_Dakota/New_Salem","America/North_Dakota/Beulah"]},"lighted":{"type":"boolean","description":"Whether the park has lights.","example":true},"bathrooms":{"type":"boolean","description":"Whether the park has bathrooms.","example":true},"tennisClub":{"type":"boolean","description":"Whether the park is a tennis club.","example":true},"ballWall":{"type":"boolean","description":"Whether the park has a ball wall.","example":true},"tennisStore":{"type":"boolean","description":"Whether the park has a tennis store.","example":true},"fee":{"type":"boolean","description":"Whether the park has a fee.","example":true},"restricted":{"type":"boolean","description":"Whether the park is restricted.","example":true},"courts":{"description":"The list of courts associated with the park.","example":[],"type":"array","items":{"$ref":"#/components/schemas/CreateCourtDto"}}},"required":["name","addressLine","city","state","zip","timezone"]}}}}